(posix) check ARCAN_DEBUG_NONANNY env to disable nanny-thread
FossilOrigin-Name: b5e2621c4cddd33489fd823240861c37801d54a429bbda30de0df06f6e367dc2

+++ platform/posix/frameserver.c
@@ -161,6 +161,19 @@ void arcan_frameserver_killchild(arcan_frameserver* src)
 	pthread_t pthr;
 	*pidptr = src->child;
 
+	static bool env_checked;
+	static bool no_nanny;
+
+	if (!env_checked)
+	{
+		env_checked = true;
+		if (getenv("ARCAN_DEBUG_NONANNY"))
+			no_nanny = true;
+	}
+
+	if (no_nanny)
+		return;
+
 	if (0 != pthread_create(&pthr, NULL, nanny_thread, (void*) pidptr)){
 		kill(src->child, SIGKILL);
 	}
